requirements.txt
fastapi[all]
uvicorn
pydantic
python-jose[cryptography]
python-dotenv